# packages/shared/openapi/openapi-generated.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Kyro API",
    "version": "1.0.0",
    "description": "Language-agnostic API platform for file storage and management.",
    "contact": {
      "name": "Kyro Support",
      "url": "https://kyro.io"
    },
    "license": {
      "name": "Proprietary"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000/api/v1",
      "description": "Development"
    },
    {
      "url": "https://api.kyro.io/api/v1",
      "description": "Production"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "apiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "message": {
                "type": "string"
              },
              "code": {
                "type": "string"
              },
              "details": {
                "type": "array",
                "items": {
                  "type": "object"
                }
              }
            }
          }
        }
      },
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "role": {
            "type": "string",
            "enum": [
              "owner",
              "admin",
              "member"
            ]
          },
          "orgId": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
...
    },
    "responses": {
      "BadRequest": {
        "description": "Invalid request body",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Authentication required",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
      "Conflict": {
        "description": "Resource already exists",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
    }
  },
  "paths": {
    "/auth/register": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Register new user and organisation",
        "x-scope": null,
        "x-body-description": "{ \"orgName\": \"Acme Inc\", \"email\": \"admin@acme.com\", \"password\": \"secure123\" }",
        "x-response-example": "{\"token\":\"eyJhbGciOiJIUzI1NiIs...\",\"user\":{\"id\":\"550e8400-e29b-41d4-a716-446655440000\",\"email\":\"admin@acme.com\",\"role\":\"owner\",\"orgId\":\"550e8400-e29b-41d4-a716-446655440000\"}}",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "email",
                  "password",
                  "orgName"
                ],
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email",
                    "description": "User email address"
                  },
                  "password": {
                    "type": "string",
                    "format": "password",
                    "minLength": 8,
                    "description": "User password (min 8 characters)"
                  },
                  "orgName": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 255,
                    "description": "Organisation name"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User registered successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "token": {
                      "type": "string",
                      "description": "JWT authentication token"
                    },
                    "user": {
                      "$ref": "#/components/schemas/User"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Login user",
        "x-scope": null,
        "x-body-description": "{ \"email\": \"admin@acme.com\", \"password\": \"secure123\" }",
        "x-response-example": "{\"token\":\"eyJhbGciOiJIUzI1NiIs...\",\"user\":{\"id\":\"550e8400-e29b-41d4-a716-446655440000\",\"email\":\"admin@acme.com\",\"role\":\"owner\",\"orgId\":\"550e8400-e29b-41d4-a716-446655440000\"}}",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "email",
                  "password"
                ],
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email"
                  },
                  "password": {
                    "type": "string",
                    "format": "password"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "token": {
                      "type": "string"
                    },
                    "user": {
                      "$ref": "#/components/schemas/User"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/auth/me": {
      "get": {
        "tags": [
          "Authentication"
        ],
        "summary": "Get current user",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-scope": null,
        "x-body-description": null,
        "x-response-example": "{\"id\":\"550e8400-e29b-41d4-a716-446655440000\",\"email\":\"admin@acme.com\",\"role\":\"owner\",\"orgId\":\"550e8400-e29b-41d4-a716-446655440000\"}",
        "responses": {
          "200": {
            "description": "Current user details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
...
  }
}